Adding logic to validate pools for limit before expansion
