+++ pkg/units/units.go
@@ -8,6 +8,7 @@ const (
 	MiB
 	// GiB 1024 MiB
 	GiB
+
 	// TiB 1024 GiB
 	TiB
 	// PiB 1024 TiB
